Updates to reorganize project structure
- Move files into pkg/authenticator
- Move main.go into cmd/authenticator
- Move binaries into bin/ and update Jenkinsfile
- Update build / dev scripts to account for new structure
- Move functions relating to Login / Authenticate out of main.go and into
  authenticator package
- Add (some) comments
- Move utils into authenticator.go, as those functions were only used there
