+++ bin/build
@@ -14,7 +14,7 @@ echo "building conjur-authn-k8s-client-go with tag ${TAG} <<"
 docker build -t conjur-authn-k8s-client-go:builder .
 docker run --rm -v $PWD:/go/src/github.com/cyberark/conjur-authn-k8s-client \
   conjur-authn-k8s-client-go:builder env CGO_ENABLED=0 GOOS=linux \
-  go build -a -installsuffix cgo -o authenticator .
+  go build -a -installsuffix cgo -o authenticator ./cmd/authenticator
 docker build -f Dockerfile.scratch -t "${TAG}" .
 docker build --build-arg VERSION="$VERSION" -f Dockerfile.redhat -t "${TAG}-redhat" .
