+++ pkg/authenticator/requests.go
@@ -1,11 +1,12 @@
-package main
+package authenticator
 
 import (
-	"net/http"
-	"fmt"
 	"bytes"
-	"net/url"
+	"fmt"
 	"io/ioutil"
+	"log"
+	"net/http"
+	"net/url"
 )
 
 func LoginRequest(authnURL string, conjurVersion string, csrBytes []byte) (*http.Request, error) {
@@ -17,8 +18,8 @@ func LoginRequest(authnURL string, conjurVersion string, csrBytes []byte) (*http
 		authenticateURL = fmt.Sprintf("%s/inject_client_cert", authnURL)
 	}
 
-	infoLogger.Printf("making login request to %s", authenticateURL)
-	
+	log.Printf("making login request to %s", authenticateURL)
+
 	req, err := http.NewRequest("POST", authenticateURL, bytes.NewBuffer(csrBytes))
 	if err != nil {
 		return nil, err
@@ -30,24 +31,25 @@ func LoginRequest(authnURL string, conjurVersion string, csrBytes []byte) (*http
 
 func AuthenticateRequest(authnURL string, conjurVersion string, account string, username string, cert []byte) (*http.Request, error) {
 	var authenticateURL string
-	
+
 	if conjurVersion == "4" {
 		authenticateURL = fmt.Sprintf("%s/users/%s/authenticate", authnURL, url.QueryEscape(username))
 	} else if conjurVersion == "5" {
 		authenticateURL = fmt.Sprintf("%s/%s/%s/authenticate", authnURL, account, url.QueryEscape(username))
 	}
 
-	infoLogger.Printf("making authn request to %s", authenticateURL)
-	
 	var req *http.Request
 	var err error
-	
+
+	infoLogger.Printf("making authn request to %s", authenticateURL)
+
 	if conjurVersion == "5" {
 		body := bytes.NewReader(cert)
 		req, err = http.NewRequest("POST", authenticateURL, body)
 	} else {
 		req, err = http.NewRequest("POST", authenticateURL, nil)
 	}
+
 
 	if err != nil {
 		return nil, err
@@ -56,7 +58,6 @@ func AuthenticateRequest(authnURL string, conjurVersion string, account string, 
 
 	return req, nil
 }
-
 
 func readBody(resp *http.Response) ([]byte, error) {
 	defer resp.Body.Close()
